1，热更新总会下载zip包问题 2，修复stageData会被释放问题

+++ proj/extensions/fix_engine/engine/jsb_extension_auto.cpp
@@ -2599,7 +2599,7 @@ static bool js_cc_extension_AssetsManager_setDownloadAgainZip(se::State& s)
     const auto& args = s.args();
     size_t argc = args.size();
     cc::extension::AssetsManagerEx *arg1 = (cc::extension::AssetsManagerEx *) NULL;
-    int arg2;
+    float arg2;
 
     if (argc != 1) {
         SE_REPORT_ERROR("wrong number of arguments: %d, was expecting %d", (int)argc, 1);
